requirements.txt
Flask==1.1.1
werkzeug==0.16.0

# Application server for both development and production.
gunicorn==20.0.4

# Testing and static analysis.
pytest==5.3.2
pytest-cov==2.8.1
mock==3.0.5
flake8==3.7.9

# Data and workers.
psycopg2-binary==2.8.4
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.12
alembic==1.3.2

redis==3.3.11
celery==4.4.0

# Forms.
Flask-WTF==0.14.2
WTForms-Components==0.10.4
WTForms-Alchemy==0.16.9

# Payments.
stripe==2.41.0

# Utils.
faker==3.0.0

# Extensions.
flask-debugtoolbar==0.10.1
Flask-Mail==0.9.1
Flask-Login==0.4.1
Flask-Limiter==1.1.0
Flask-Babel==0.12.2
Flask-Static-Digest==0.1.2
